requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
dataclasses-json==0.6.6
Deprecated==1.2.14
dune_client==1.7.3
frozenlist==1.4.1
idna==3.7
marshmallow==3.21.2
multidict==6.0.5
mypy-extensions==1.0.0
ndjson==0.3.1
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pydantic==1.10.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.29.0
shroomdk==2.0.8
six==1.16.0
types-Deprecated==1.2.9.20240311
types-python-dateutil==2.9.0.20240316
types-PyYAML==6.0.12.20240311
types-requests==2.31.0.6
types-setuptools==70.0.0.20240524
types-urllib3==1.26.25.14
typing-inspect==0.9.0
typing_extensions==4.12.0
tzdata==2024.1
urllib3==1.26.18
wrapt==1.16.0
yarl==1.9.4
